Update fr.po
C'est tout!

I actually don't know, it's the first time for me on github: I saw that there is in the file a "language-team". Do I have to be part of one of them to edit files? While my Italian is top notch because it's my mother tongue, French and Spanish are aquired languages, and I am qualified enough to edit files but not much to have proper discussions in some language group, if you know what I mean. 

All the best, 
Luca